+++ src/Services/Catalog/ProductImage/Result/ProductImagesResult.php
@@ -3,7 +3,7 @@
 /**
  * This file is part of the bitrix24-php-sdk package.
  *
- * © Maksim Mesilov <mesilov.maxim@gmail.com>
+ * © Dmitriy Ignatenko <algonexys@gmail.com>
  *
  * For the full copyright and license information, please view the MIT-LICENSE.txt
  * file that was distributed with this source code.
@@ -11,22 +11,22 @@
 
 declare(strict_types=1);
 
-namespace Bitrix24\SDK\Services\Catalog\Product\Result;
+namespace Bitrix24\SDK\Services\Catalog\ProductImage\Result;
 
 use Bitrix24\SDK\Core\Exceptions\BaseException;
 use Bitrix24\SDK\Core\Result\AbstractResult;
 
-class ProductsResult extends AbstractResult
+class ProductImagesResult extends AbstractResult
 {
     /**
-     * @return ProductItemResult[]
+     * @return ProductImageItemResult[]
      * @throws BaseException
      */
-    public function getProducts(): array
+    public function getProductImages(): array
     {
         $res = [];
-        foreach ($this->getCoreResponse()->getResponseData()->getResult()['products'] as $product) {
-            $res[] = new ProductItemResult($product);
+        foreach ($this->getCoreResponse()->getResponseData()->getResult()['productImages'] as $productImage) {
+            $res[] = new ProductImageItemResult($productImage);
         }
 
         return $res;